style: Implement responsive layout and styles for smaller screens
Add responsive styles using media queries for screens with a maximum width of 900px.

For the .header class, I adjusted the flex-direction property to column, resulting in a vertical stacking of the header elements. I also introduced a gap of 20px between the stacked elements.

Within the .header-menu a class, I applied a semi-transparent background color with an RGBA value of (250, 235, 215, 0.651). Additionally, I added a border-radius of 4px and adjusted the padding to provide comfortable spacing within the menu links.